feat: Add Map and Lineup models!

+++ src/main/java/com/helper/vavahelper/Repositories/SkillsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.helper.vavahelper.models.Skills.Skills;
 
-public interface SkillsRepository extends JpaRepository<Skills, Integer>{
+public interface SkillsRepository extends JpaRepository<Skills, Long>{
     
 }
